fixing ui test case.
