+++ ui/src/app/pages/data/data-set/dataset-page.spec.tsx
@@ -7,7 +7,6 @@ import {dataSetApi, registerApiClient} from 'app/services/swagger-fetch-clients'
 import {currentWorkspaceStore, NavStore, urlParamsStore} from 'app/utils/navigation';
 import {
   CohortsApi,
-  ConceptsApi,
   ConceptSetsApi,
   DataSetApi,
   WorkspaceAccessLevel
@@ -15,7 +14,6 @@ import {
 import {waitOneTickAndUpdate} from 'testing/react-test-helpers';
 import {CohortsApiStub, exampleCohortStubs} from 'testing/stubs/cohorts-api-stub';
 import {ConceptSetsApiStub} from 'testing/stubs/concept-sets-api-stub';
-import {ConceptsApiStub} from 'testing/stubs/concepts-api-stub';
 import {DataSetApiStub} from 'testing/stubs/data-set-api-stub';
 import {
   workspaceDataStub,
@@ -26,7 +24,6 @@ import {
 describe('DataSetPage', () => {
   beforeEach(() => {
     registerApiClient(CohortsApi, new CohortsApiStub());
-    registerApiClient(ConceptsApi, new ConceptsApiStub());
     registerApiClient(ConceptSetsApi, new ConceptSetsApiStub());
     registerApiClient(DataSetApi, new DataSetApiStub());
     urlParamsStore.next({
