+++ ui/src/testing/stubs/concepts-api-stub.ts
@@ -222,24 +222,4 @@ export class ConceptsApiStub extends ConceptsApi {
     });
   }
 
-  public getValuesFromDomain(workspaceNamespace: string, workspaceId: string, domain: string)
-  : Promise<DomainValuesResponse> {
-    const domainValueItems = [];
-    switch (domain) {
-      case 'CONDITION':
-        domainValueItems.push({value: 'Condition1'});
-        domainValueItems.push({value: 'Condition2'});
-        break;
-      case 'MEASUREMENT':
-        domainValueItems.push({value: 'Measurement1'});
-        domainValueItems.push({value: 'Measurement2'});
-        domainValueItems.push({value: 'Measurement3'});
-        break;
-      case 'DRUG':
-        domainValueItems.push({value: 'Drug1'});
-        break;
-    }
-    return Promise.resolve({items: domainValueItems});
-  }
-
 }
